Create escape utility script and fix axois baseURL
The escapeFileURL function defined in the new escape utility script
helps escape characters like # in file names or directories. This is
important for links to work properly.

Also, the baseURL for the browser instance of axios now has the route
appended so that different instances can use other functions in the
backend in the future.